Move referências a divisões e núcleos para do app divisoes para o core
