+++ sme_coad_apps/core/api/viewsets/divisao_viewset.py
@@ -1,6 +1,6 @@
 from rest_framework import viewsets
 
-from sme_coad_apps.divisoes.models import Divisao
+from ...models import Divisao
 from ..serializers.divisao_serializer import DivisaoSerializer
