+++ sme_coad_apps/core/tests/test_nucleo_endpoints.py
